Move buttonInjector folder into contentScript folder
