+++ src/contentScript/buttonInjector/ButtonInjectorFactory.ts
